azure_function/requirements.txt
pandas
scikit-learn
azure-functions